SkinnedModel: Finish populating rig data code

+++ src/Rig.h
@@ -1,5 +1,4 @@
 #pragma once
-#include "AK/Math.h"
 #include "AK/Types.h"
 
 #include <string>
@@ -13,7 +12,7 @@ struct xform;
 struct Rig
 {
     std::vector<std::string> bone_names = {};
-    std::vector<i16> parents = {};
+    std::vector<i32> parents = {};
     std::vector<AK::xform> ref_pose = {};
     u32 num_bones = 0;
 };
